chore: split popup-ui from desktop-ui
