+++ packages/app/src/popup-ui/html/popup-ui.html
@@ -1,0 +1,28 @@
+<!doctype html>
+<html data-testid="mmd-popup">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1 user-scalable=no">
+    <title>MetaMask Desktop</title>
+    <link rel="stylesheet" type="text/css" href="./index.css" title="ltr">
+    <link rel="stylesheet" type="text/css" href="./index-rtl.css" title="rtl" disabled>
+  </head>
+  <body>
+    <div id="mmd-popup-draggable-hidden-bar"></div>
+    <div id="mmd-popup-root"></div>
+    <script src="./sentry-install.js" type="text/javascript" charset="utf-8"></script>
+    {{@if(it.applyLavaMoat)}}
+      <script src="./runtime-lavamoat.js" type="text/javascript" charset="utf-8"></script>
+      <script src="./lockdown-more.js" type="text/javascript" charset="utf-8"></script>
+      <script src="./policy-load.js" type="text/javascript" charset="utf-8"></script>
+    {{#else}}
+      <script src="./lockdown-install.js" type="text/javascript" charset="utf-8"></script>
+      <script src="./lockdown-run.js" type="text/javascript" charset="utf-8"></script>
+      <script src="./lockdown-more.js" type="text/javascript" charset="utf-8"></script>
+      <script src="./runtime-cjs.js" type="text/javascript" charset="utf-8"></script>
+    {{/if}}
+    {{@each(it.jsBundles) => val}}
+      <script src="{{val}}" type="text/javascript" charset="utf-8"></script>
+    {{/each}}
+  </body>
+</html>
