+++ packages/app/src/popup-ui/pages/routes/index.js
@@ -1,0 +1,1 @@
+export { default } from './routes.container';
