+++ packages/app/src/popup-ui/pages/routes/routes.container.js
@@ -1,0 +1,42 @@
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+import { compose } from 'redux';
+
+import {
+  getFirstPermissionRequest,
+  getFirstSnapInstallOrUpdateRequest,
+  getUnapprovedTemplatedConfirmations,
+  unconfirmedTransactionsCountSelector,
+} from '../../../../submodules/extension/ui/selectors';
+import Routes from './routes.component';
+
+function getPermissionRequestId(state) {
+  const permissionRequest = getFirstPermissionRequest(state);
+  let permissionRequestId = permissionRequest?.metadata?.id;
+
+  if (!permissionRequestId) {
+    const snapRequest = getFirstSnapInstallOrUpdateRequest(state);
+    permissionRequestId = snapRequest?.metadata?.id;
+  }
+
+  return permissionRequestId;
+}
+
+function mapStateToProps(state) {
+  const stateLoaded = state.metamask.desktopEnabled;
+
+  const permissionRequestId = stateLoaded
+    ? getPermissionRequestId(state)
+    : undefined;
+
+  return {
+    hasPendingApproval:
+      stateLoaded && getUnapprovedTemplatedConfirmations(state).length > 0,
+    hasPendingPermissionApproval: Boolean(permissionRequestId),
+    hasPendingTransactionApproval:
+      stateLoaded && unconfirmedTransactionsCountSelector(state) > 0,
+    permissionRequestId,
+  };
+}
+
+export default compose(withRouter, connect(mapStateToProps))(Routes);
