+++ packages/app/src/popup-ui/store/store.js
@@ -1,0 +1,8 @@
+import { createStore, applyMiddleware } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+
+export default function configureStore(initialState = {}) {
+  const storeEnhancers = applyMiddleware(thunkMiddleware);
+  const store = createStore(initialState, storeEnhancers);
+  return { store };
+}
