+++ packages/app/src/ui/ducks/index.js
@@ -3,14 +3,6 @@ import { persistReducer, createMigrate } from 'redux-persist';
 
 import pairStatusMigrations from '../migrations/pair-status';
 import rootMigrations from '../migrations/root';
-import sendReducer from '../../../submodules/extension/ui/ducks/send/send';
-import historyReducer from '../../../submodules/extension/ui/ducks/history/history';
-import appStateReducer from '../../../submodules/extension/ui/ducks/app/app';
-import confirmTransactionReducer from '../../../submodules/extension/ui/ducks/confirm-transaction/confirm-transaction.duck';
-import metamaskReducer from '../../../submodules/extension/ui/ducks/metamask/metamask';
-import domainsReducer from '../../../submodules/extension/ui/ducks/domains';
-import gasReducer from '../../../submodules/extension/ui/ducks/gas/gas.duck';
-import localeMessagesReducer from '../../../submodules/extension/ui/ducks/locale/locale';
 import appReducer from './app/app';
 import pairStatusReducer from './pair-status/pair-status';
 
@@ -38,14 +30,6 @@ const persistedAppReducer = persistReducer(appPersistConfig, appReducer);
 const rootReducer = combineReducers({
   app: persistedAppReducer,
   pairStatus: persistedPairStatusReducer,
-  metamask: metamaskReducer,
-  localeMessages: localeMessagesReducer,
-  send: sendReducer,
-  history: historyReducer,
-  appState: appStateReducer,
-  confirmTransaction: confirmTransactionReducer,
-  DNS: domainsReducer,
-  gas: gasReducer,
 });
 
 export default rootReducer;
