+++ packages/app/src/ui/index.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import { _setBackgroundConnection } from '../../submodules/extension/ui/store/action-queue';
-import { IPCMainStream } from '../app/ipc-main-stream';
-import metaRPCClientFactory from '../../submodules/extension/app/scripts/lib/metaRPCClientFactory';
-import * as actions from './actions';
 import configureStore from './store/store';
 import Root from './pages';
 import registerUpdatePairStatus from './hooks/registerUpdatePairStatus';
@@ -34,20 +30,6 @@ async function launchDesktopUi() {
   // Register window event listeners from the main process
   window.electronBridge.onResized(registerResizedEvent(store));
   window.electronBridge.onMoved(registerMovedEvent(store));
-
-  window.electronBridge.addBackgroundMessageListener(async (data) => {
-    if (data.data?.method !== 'sendUpdate') {
-      return;
-    }
-
-    store.dispatch(actions.updateState(data.data?.params[0]));
-  });
-
-  const backgroundStream = metaRPCClientFactory(
-    new IPCMainStream(window.electronBridge),
-  );
-
-  _setBackgroundConnection(backgroundStream);
 
   render(
     <Root store={store} persistor={persistor} />,
