+++ packages/app/src/ui/pages/index.js
@@ -7,14 +7,6 @@ import PropTypes from 'prop-types';
 import * as Sentry from '@sentry/electron/renderer';
 import { I18nProvider } from '../contexts/i18n';
 import { EVENT_NAMES } from '../../app/metrics/metrics-constants';
-import {
-  I18nProvider as ExtensionI18nProvider,
-  LegacyI18nProvider,
-} from '../../../submodules/extension/ui/contexts/i18n';
-import {
-  MetaMetricsProvider,
-  LegacyMetaMetricsProvider,
-} from '../../../submodules/extension/ui/contexts/metametrics';
 import Routes from './routes';
 import CriticalError from './error/critical-error.component';
 
@@ -49,17 +41,9 @@ class Root extends PureComponent {
       <Provider store={store}>
         <PersistGate loading={null} persistor={persistor}>
           <HashRouter>
-            <MetaMetricsProvider>
-              <LegacyMetaMetricsProvider>
-                <I18nProvider>
-                  <ExtensionI18nProvider>
-                    <LegacyI18nProvider>
-                      <Routes />
-                    </LegacyI18nProvider>
-                  </ExtensionI18nProvider>
-                </I18nProvider>
-              </LegacyMetaMetricsProvider>
-            </MetaMetricsProvider>
+            <I18nProvider>
+              <Routes />
+            </I18nProvider>
           </HashRouter>
         </PersistGate>
       </Provider>
